Dont use rAsm implicitly in x64Assembler
This removes the implicit uses of rAsm from X64Assembler,
and moves the definition to abi-x64.h.  Some mentions of
rAsm were hoisted, but most involved doing stores, so those
were converted to store two 32-bit immediates (without
using a scratch register) when possible.

The Immed wrapper class now only allows 32-bit immediates,
and a new Immed64 wrapper is available for immediates whose
compile-time-type (vm-compile time) is 64 bits.

Reviewed By: @ottoni

Differential Revision: D1224706

+++ hphp/runtime/base/ref-data.h
@@ -127,7 +127,7 @@ struct RefData {
     not_reached();
 #endif
   }
-  static constexpr ptrdiff_t tvOffset() { return offsetof(RefData, m_tv); }
+  static constexpr int tvOffset() { return offsetof(RefData, m_tv); }
 
   void assertValid() const {
     assert(m_magic == Magic::kMagic);
